localStorage storing temp states. Roadblock in createClient()
Stumbled upon a hiccup, turns out, in a Nextjs Project with GELDATA, if we use createClient() in a client component, it will return an error since it's only supposed to be used on server components. They do provide an alternative, createHTTPClient(), however, I believe it's best to stick server operations on the server. Thus, I will resort to a safer api routes system. Please refer to the last chat inquiry in the Grok superego chat.